refactor: move schemas to a new file and simplify collection definitions
